Introduce first version of makedeb package
This change introduces the first version of the makedeb package. The
package allows for creating .deb packages, a file format used for
managing software packages by various Linux distributions such as Debian
and Ubuntu. The package's goal is solely to provide the functionality
for creating packages. It does not prescribe the resulting layout of
files in the file system, when such a package is installed. This
knowledge is generally distribution specific and needs to be handled in
a per package basis.