Ensure destination directory exists when copying content
The copy2 function used for copying single files cannot deal with cases
where the path to the destination file does not yet exist. That is, it
will not create missing directories in this path.
This change adjusts the function responsible for copying the content to
ensure that the destination directory exists before invoking copy2.