Add heuristic for inferring proper access mode for content
We have encountered a case where a file was packaged with permissions
that were not suitable for later installation: the file was created
temporarily and, hence, was only user accessible.
This case highlights a deeper problem in that we do not adjust the file
permissions at all (we do adjust the owner/group, though. The perhaps
most comprehensive fix would ask the user for a specification of the
desired access mode. However, that requires additional data from the
user and having a nice specification format is not straight forward: the
user typically wants to package entire directory structures and it is
not clear how to specify the different modes typically used for regular
files and directories.
To that end, this change adds a heuristic that takes care of
automatically inferring the proper access rights. The idea two fold.
First, directories will always have 755 access. For regular files we
assume 644 access, but if one of the (user, group, other) executable
bits is set we up that to 755 as well.