Add support for adding custom DEBIAN files
The .deb file format supports additional control files that influence
the behavior of the package. E.g., a postinst file can be included which
is a script that is executed after the installation of a package. More
such files with special meaning exist and can be found in the
documentation of the file format. All these files are to be included in
the control tarball.
This change adds support for adding such files to the .deb file we
produce.
The interface introduced allows for passing in a list of files to be
used as such custom control files. The rationale for not providing a
per-file interface is that writing such files requires intimate
knowledge of the .deb format. With such knowledge being required we can
also assume that users have knowledge of the names of those files. The
alternative would be to pass in the individual files directly (e.g.,
postinst=None) but the fear is that this is not very future proof while
also requiring many more keyword arguments.