Ensure to add all data directories
We have seen a case where a created package failed to install because a
directory a file was supposed to be installed in did not exist. The
underlying problem is that the data archive did not explicitly contain
the directory but only the file in it. E.g., data.tar.xz contains:
  > etc/program/file

When really it should contain the following:
  > etc
  > etc/program
  > etc/program/file

That is, all contained directories need to be listed explicitly. This
change adjusts the package to adhere to this rule. To implement this
functionality a slight refactoring had to be performed: Instead of
having a package directory containing the control as well as data files
we now create two separate directories, one for data and one for control
files.